return special exp values to ints, per discussion with Zacam awhile ago while we're at it, change kamikaze damage to int since it's treated as such

+++ code/fred2/missionsave.cpp
@@ -1686,24 +1686,24 @@ int CFred_mission_save::save_objects()
 
 					required_string_fred("+Special Exp Damage:"); 
 					parse_comments();
-					fout(" %f", shipp->special_exp_damage);
+					fout(" %d", shipp->special_exp_damage);
 
 					required_string_fred("+Special Exp Blast:"); 
 					parse_comments();
-					fout(" %f", shipp->special_exp_blast);
+					fout(" %d", shipp->special_exp_blast);
 
 					required_string_fred("+Special Exp Inner Radius:"); 
 					parse_comments();
-					fout(" %f", shipp->special_exp_inner);
+					fout(" %d", shipp->special_exp_inner);
 
 					required_string_fred("+Special Exp Outer Radius:"); 
 					parse_comments();
-					fout(" %f", shipp->special_exp_outer);
+					fout(" %d", shipp->special_exp_outer);
 
 					if (shipp->use_shockwave && (shipp->special_exp_shockwave_speed > 0)) {
 						optional_string_fred("+Special Exp Shockwave Speed:"); 
 						parse_comments();
-						fout(" %f", shipp->special_exp_shockwave_speed);
+						fout(" %d", shipp->special_exp_shockwave_speed);
 					}
 					else {
 						bypass_comment(";;FSO 3.6.13;; +Special Exp Shockwave Speed:");
@@ -1723,20 +1723,20 @@ int CFred_mission_save::save_objects()
 					fout_version("\n$Special Explosion:");
 
 					fout_version("\n+Special Exp Damage:"); 
-					fout(" %f", shipp->special_exp_damage);
+					fout(" %d", shipp->special_exp_damage);
 
 					fout_version("\n+Special Exp Blast:"); 
-					fout(" %f", shipp->special_exp_blast);
+					fout(" %d", shipp->special_exp_blast);
 
 					fout_version("\n+Special Exp Inner Radius:"); 
-					fout(" %f", shipp->special_exp_inner);
+					fout(" %d", shipp->special_exp_inner);
 
 					fout_version("\n+Special Exp Outer Radius:"); 
-					fout(" %f", shipp->special_exp_outer);
+					fout(" %d", shipp->special_exp_outer);
 
 					if (shipp->use_shockwave && (shipp->special_exp_shockwave_speed > 0)) {
 						fout_version("\n+Special Exp Shockwave Speed:"); 
-						fout(" %f", shipp->special_exp_shockwave_speed);
+						fout(" %d", shipp->special_exp_shockwave_speed);
 					}
 
 					if (shipp->special_exp_deathroll_time > 0) {
@@ -1815,7 +1815,7 @@ int CFred_mission_save::save_objects()
 				fout("\n+Kamikaze Damage:");
 			}
 
-			fout(" %d", (int)(Ai_info[shipp->ai_index].kamikaze_damage) );
+			fout(" %d", Ai_info[shipp->ai_index].kamikaze_damage);
 		}
 
 		if (shipp->hotkey != -1) {
